move Item component and remove FC
